Fix references to moved file
