+++ tests/unit_tests/hyperion/device_setup_plans/test_manipulate_sample.py
@@ -5,7 +5,7 @@ from bluesky.run_engine import RunEngine
 from dodal.devices.aperturescatterguard import ApertureScatterguard, ApertureValue
 from ophyd_async.testing import get_mock_put
 
-from mx_bluesky.hyperion.device_setup_plans.manipulate_sample import (
+from mx_bluesky.common.device_setup_plans.manipulate_sample import (
     move_aperture_if_required,
     move_phi_chi_omega,
     move_x_y_z,
